refactor: change category file structure
