+++ src/feature/catalog/categories/use-category-tree.ts
@@ -1,0 +1,11 @@
+import { arrayToTree } from 'performant-array-to-tree';
+
+import type { BasicCategory, CategoryNode } from '../types';
+
+export const useCategoryTree = (categories?: BasicCategory[]): CategoryNode[] => {
+  if (!categories) return [];
+  return arrayToTree(categories, {
+    parentId: 'parent.obj.id',
+    dataField: null,
+  }) as CategoryNode[];
+};
