+++ src/pages/catalog-page.tsx
@@ -4,7 +4,7 @@ import useSWR from 'swr';
 import { Spinner } from '@/components/ui/spiner';
 import { fetchProducts } from '@/feature/catalog/api/fetch-products';
 import { Breadcrumbs } from '@/feature/catalog/breadcrumbs';
-import { CategoryNavigation } from '@/feature/catalog/category-navigation';
+import { CategoryNavigation } from '@/feature/catalog/categories/category-navigation';
 import { ProductList } from '@/feature/catalog/product-list';
 import { useCategoryStore } from '@/service/store/use-category-store';
